src/bee_detection_model/requirements.txt
Pillow
fastapi
uvicorn
numpy
tensorflow
tensorflow_hub
python-multipart
ssl